Finish html design and css style, update javsscr.
Added images to buttons
Added counters on html page as well as an output field for results of
round
finished javaScript, i think.
still need to see why it's not working
might be the button click event
could not be done correctly? and therefore does not get registered when
i want it to?
re-check game logic...